Rename `branchStoresCache.ts` -> `virtualBranch.ts`
- split out `baseBranch.ts`, these should never have been in the same file
