+++ app/src/lib/vbranches/baseBranch.ts
@@ -1,5 +1,5 @@
-import { BaseBranch, Branch, VirtualBranches } from './types';
-import { Code, invoke, listen } from '$lib/backend/ipc';
+import { BaseBranch } from './types';
+import { Code, invoke } from '$lib/backend/ipc';
 import { observableToStore } from '$lib/rxjs/store';
 import * as toasts from '$lib/utils/toasts';
 import { plainToInstance } from 'class-transformer';
@@ -10,107 +10,13 @@ import {
 	catchError,
 	BehaviorSubject,
 	debounceTime,
-	concat,
-	from,
 	tap,
-	map,
-	firstValueFrom,
-	timeout,
 	combineLatest,
-	of,
 	startWith,
 	Subject,
 	mergeWith
 } from 'rxjs';
-import { writable, type Readable } from 'svelte/store';
-
-export class VirtualBranchService {
-	branches$: Observable<Branch[] | undefined>;
-	stashedBranches$: Observable<Branch[] | undefined>;
-	activeBranches$: Observable<Branch[] | undefined>;
-	branchesError = writable<any>();
-	private reload$ = new BehaviorSubject<void>(undefined);
-	private fresh$ = new Subject<void>();
-
-	activeBranches: Readable<Branch[] | undefined>;
-	activeBranchesError: Readable<any>;
-
-	constructor(
-		private projectId: string,
-		gbBranchActive$: Observable<boolean>
-	) {
-		this.branches$ = this.reload$.pipe(
-			switchMap(() => gbBranchActive$),
-			switchMap((gbBranchActive) =>
-				gbBranchActive
-					? concat(
-							from(listVirtualBranches({ projectId })),
-							new Observable<Branch[]>((subscriber) => {
-								return subscribeToVirtualBranches(projectId, (branches) =>
-									subscriber.next(branches)
-								);
-							})
-						)
-					: of([])
-			),
-			tap((branches) => {
-				branches.forEach((branch) => {
-					branch.files.sort((a) => (a.conflicted ? -1 : 0));
-					branch.isMergeable = invoke<boolean>('can_apply_virtual_branch', {
-						projectId: projectId,
-						branchId: branch.id
-					});
-				});
-				this.fresh$.next(); // Notification for fresh reload
-			}),
-			startWith(undefined),
-			shareReplay(1)
-		);
-
-		this.stashedBranches$ = this.branches$.pipe(
-			map((branches) => branches?.filter((b) => !b.active))
-		);
-
-		this.activeBranches$ = this.branches$.pipe(
-			map((branches) => branches?.filter((b) => b.active))
-		);
-
-		[this.activeBranches, this.activeBranchesError] = observableToStore(this.activeBranches$);
-	}
-
-	async reload() {
-		this.branchesError.set(undefined);
-		const fresh = firstValueFrom(
-			this.fresh$.pipe(
-				timeout(30000),
-				catchError(() => {
-					// Observable never errors for any other reasons
-					const err = 'Timed out while reloading virtual branches';
-					console.warn(err);
-					toasts.error(err);
-					return of();
-				})
-			)
-		);
-		this.reload$.next();
-		return await fresh;
-	}
-
-	async getById(branchId: string) {
-		return await firstValueFrom(
-			this.branches$.pipe(
-				timeout(10000),
-				map((branches) => branches?.find((b) => b.id == branchId && b.upstream))
-			)
-		);
-	}
-}
-
-function subscribeToVirtualBranches(projectId: string, callback: (branches: Branch[]) => void) {
-	return listen<any>(`project://${projectId}/virtual-branches`, (event) =>
-		callback(plainToInstance(VirtualBranches, event.payload).branches)
-	);
-}
+import type { Readable } from 'svelte/store';
 
 export class NoDefaultTarget extends Error {}
 
@@ -193,9 +99,8 @@ export class BaseBranchService {
 	}
 }
 
-export async function listVirtualBranches(params: { projectId: string }): Promise<Branch[]> {
-	return plainToInstance(VirtualBranches, await invoke<any>('list_virtual_branches', params))
-		.branches;
+async function getBaseBranch(params: { projectId: string }): Promise<BaseBranch | null> {
+	return plainToInstance(BaseBranch, await invoke<any>('get_base_branch_data', params));
 }
 
 export async function getRemoteBranches(projectId: string | undefined) {
@@ -206,8 +111,4 @@ export async function getRemoteBranches(projectId: string | undefined) {
 			.sort((a, b) => a.localeCompare(b))
 			.map((name) => ({ name }))
 	);
-}
-
-async function getBaseBranch(params: { projectId: string }): Promise<BaseBranch | null> {
-	return plainToInstance(BaseBranch, await invoke<any>('get_base_branch_data', params));
 }
