+++ app/src/lib/vbranches/branchController.ts
@@ -3,8 +3,9 @@ import { showToast } from '$lib/notifications/toasts';
 import * as toasts from '$lib/utils/toasts';
 import posthog from 'posthog-js';
 import type { RemoteBranchService } from '$lib/stores/remoteBranches';
-import type { BaseBranchService, VirtualBranchService } from './branchStoresCache';
+import type { BaseBranchService } from './baseBranch';
 import type { Branch, Hunk, LocalFile } from './types';
+import type { VirtualBranchService } from './virtualBranch';
 
 export class BranchController {
 	constructor(
