+++ app/src/routes/[projectId]/+layout.ts
@@ -2,8 +2,9 @@ import { BranchService } from '$lib/branches/service';
 import { getFetchNotifications } from '$lib/stores/fetches';
 import { getHeads } from '$lib/stores/head';
 import { RemoteBranchService } from '$lib/stores/remoteBranches';
+import { BaseBranchService } from '$lib/vbranches/baseBranch';
 import { BranchController } from '$lib/vbranches/branchController';
-import { BaseBranchService, VirtualBranchService } from '$lib/vbranches/branchStoresCache';
+import { VirtualBranchService } from '$lib/vbranches/virtualBranch';
 import { error } from '@sveltejs/kit';
 import { map } from 'rxjs';
 import type { Project } from '$lib/backend/projects';
